Refactor arguments conversion
This fixes a bug with converting inputs with list containing
complex types
