+++ RELEASE.md
@@ -1,0 +1,4 @@
+Release type: patch
+
+This releases fixes an issue when converting complex input types, 
+now it should support lists of complex types properly.
